fix: move common exceptions to own package
